requirements.txt
asgiref==3.4.1
boto3==1.18.48
botocore==1.21.48
Django==3.2.7
django-storages==1.11.1
djangorestframework==3.12.4
djrest-wrapper==0.3.0
drf-writable-nested==0.6.3
gunicorn==20.1.0
jmespath==0.10.0
PyJWT==2.1.0
python-dateutil==2.8.2
python-magic==0.4.24
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.1
